workflow-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
redis==5.0.1
python-dotenv==1.0.0
pyjwt>=2.8.0,<3.0.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
python-multipart==0.0.6
httpx==0.25.2
pyyaml==6.0.1
jinja2==3.1.2
structlog==23.2.0
loguru==0.7.2